chore: split content normalisation out
