+++ packages/ipfs-core-utils/package.json
@@ -40,5 +40,8 @@
     "delay": "^4.3.0",
     "dirty-chai": "^2.0.1",
     "it-all": "^1.0.1"
+  },
+  "browser": {
+    "./src/files/normalise-content": "./src/files/normalise-content.browser.js"
   }
 }
