+++ packages/ipfs-core-utils/src/files/normalise-input.js
@@ -3,6 +3,7 @@
 const errCode = require('err-code')
 const { Buffer } = require('buffer')
 const globalThis = require('ipfs-utils/src/globalthis')
+const normaliseContent = require('./normalise-content')
 
 /*
  * Transform one of:
@@ -37,10 +38,17 @@ const globalThis = require('ipfs-utils/src/globalthis')
  * AsyncIterable<{ path, content: Iterable<Bytes> }> [multiple files]
  * AsyncIterable<{ path, content: AsyncIterable<Bytes> }> [multiple files]
  * ```
- * Into:
+ *
+ * In node:
  *
  * ```
- * AsyncIterable<{ path, content: AsyncIterable<Buffer> }>
+ * AsyncIterable<{ path, mode, mtime, content: AsyncIterable<Buffer> }>
+ * ```
+ *
+ * And in the browser:
+ *
+ * ```
+ * AsyncIterable<{ path, mode, mtime, content: Blob }>
  * ```
  *
  * @param input Object
@@ -151,7 +159,7 @@ module.exports = function normaliseInput (input) {
   throw errCode(new Error('Unexpected input: ' + typeof input), 'ERR_UNEXPECTED_INPUT')
 }
 
-function toFileObject (input) {
+async function toFileObject (input) {
   const obj = {
     path: input.path || '',
     mode: input.mode,
@@ -159,75 +167,12 @@ function toFileObject (input) {
   }
 
   if (input.content) {
-    obj.content = toAsyncIterable(input.content)
+    obj.content = await normaliseContent(input.content)
   } else if (!input.path) { // Not already a file object with path or content prop
-    obj.content = toAsyncIterable(input)
+    obj.content = await normaliseContent(input)
   }
 
   return obj
-}
-
-function toAsyncIterable (input) {
-  // Bytes | String
-  if (isBytes(input) || typeof input === 'string') {
-    return (async function * () { // eslint-disable-line require-await
-      yield toBuffer(input)
-    })()
-  }
-
-  // Bloby
-  if (isBloby(input)) {
-    return blobToAsyncGenerator(input)
-  }
-
-  // Browser stream
-  if (typeof input.getReader === 'function') {
-    return browserStreamToIt(input)
-  }
-
-  // Iterator<?>
-  if (input[Symbol.iterator]) {
-    return (async function * () { // eslint-disable-line require-await
-      const iterator = input[Symbol.iterator]()
-      const first = iterator.next()
-      if (first.done) return iterator
-
-      // Iterable<Number>
-      if (Number.isInteger(first.value)) {
-        yield toBuffer(Array.from((function * () {
-          yield first.value
-          yield * iterator
-        })()))
-        return
-      }
-
-      // Iterable<Bytes>
-      if (isBytes(first.value)) {
-        yield toBuffer(first.value)
-        for (const chunk of iterator) {
-          yield toBuffer(chunk)
-        }
-        return
-      }
-
-      throw errCode(new Error('Unexpected input: ' + typeof input), 'ERR_UNEXPECTED_INPUT')
-    })()
-  }
-
-  // AsyncIterable<Bytes>
-  if (input[Symbol.asyncIterator]) {
-    return (async function * () {
-      for await (const chunk of input) {
-        yield toBuffer(chunk)
-      }
-    })()
-  }
-
-  throw errCode(new Error(`Unexpected input: ${input}`), 'ERR_UNEXPECTED_INPUT')
-}
-
-function toBuffer (chunk) {
-  return isBytes(chunk) ? chunk : Buffer.from(chunk)
 }
 
 function isBytes (obj) {
@@ -243,15 +188,6 @@ function isFileObject (obj) {
   return typeof obj === 'object' && (obj.path || obj.content)
 }
 
-function blobToAsyncGenerator (blob) {
-  if (typeof blob.stream === 'function') {
-    // firefox < 69 does not support blob.stream()
-    return browserStreamToIt(blob.stream())
-  }
-
-  return readBlob(blob)
-}
-
 async function * browserStreamToIt (stream) {
   const reader = stream.getReader()
 
@@ -263,36 +199,5 @@ async function * browserStreamToIt (stream) {
     }
 
     yield result.value
-  }
-}
-
-async function * readBlob (blob, options) {
-  options = options || {}
-
-  const reader = new globalThis.FileReader()
-  const chunkSize = options.chunkSize || 1024 * 1024
-  let offset = options.offset || 0
-
-  const getNextChunk = () => new Promise((resolve, reject) => {
-    reader.onloadend = e => {
-      const data = e.target.result
-      resolve(data.byteLength === 0 ? null : data)
-    }
-    reader.onerror = reject
-
-    const end = offset + chunkSize
-    const slice = blob.slice(offset, end)
-    reader.readAsArrayBuffer(slice)
-    offset = end
-  })
-
-  while (true) {
-    const data = await getNextChunk()
-
-    if (data == null) {
-      return
-    }
-
-    yield Buffer.from(data)
   }
 }
