+++ packages/ipfs-http-client/src/lib/multipart-request.browser.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const normaliseInput = require('./normalise-input.browser')
+const normaliseInput = require('ipfs-core-utils/src/files/normalise-input')
 const modeToString = require('./mode-to-string')
 const mtimeToObject = require('./mtime-to-object')
 const { File, FormData } = require('ipfs-utils/src/globalthis')
